_v1c

+++ controllers/contacts/removeById.js
@@ -5,6 +5,11 @@ const { RequestError } = require("../../helpers");
 const removeById = async (req, res) => {
     
       const { contactId } = req.params;
+      const { _id: owner} = req.user;
+      
+      const contact = await Contact.findOne({ owner, _id: contactId });
+      if (!contact) throw RequestError(404, 'Not authorized');
+
       const result = await Contact.findByIdAndRemove(contactId);
       if (!result) {
         throw RequestError(404, "Not found");
@@ -15,3 +20,5 @@ const removeById = async (req, res) => {
   }
 
 module.exports = removeById;
+
+
